==> 01_data_preprocessing.Rmd <==
---
title: "data_preprocessing"
author: "Adam Drożyński"
date: "4/2/2022"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list = ls())
library(pacman)
p_load(tictoc,tidyverse,dplyr,broom,ggplot2,clubSandwich,
       data.table,matrixStats,magrittr,MASS,R.utils,
       stargazer,DT,furrr, lfe,estimatr,boot,plm,lmtest,multiwayvcov,
       sandwich,glmnet,ISLR, parallel, haven, MatchIt, optmatch, ipw, AER, rdd)
```

```{r}
# Windows
#PATH_BHPS = '..\\UKDA-6614-stata\\stata\\stata13_se\\bhps\\'
#PATH_UKHLS = '..\\UKDA-6614-stata\\stata\\stata13_se\\ukhls\\'

# Linux
PATH_BHPS = '../UKDA-6614-stata/stata/stata13_se/bhps/'
PATH_UKHLS = '../UKDA-6614-stata/stata/stata13_se/ukhls/'

get_full_label_bhps <- function(label, letter){
  return(paste0('b', letter, '_', label))
}

get_full_label_ukhls <- function(label, letter){
  return(paste0(letter, '_', label))
}

get_data_path_bhps <- function(label, letter){
  return(paste0(PATH_BHPS, 'b', letter, '_', label,'.dta'))
}

get_data_path_ukhls <- function(label, letter){
  return(paste0(PATH_UKHLS, letter, '_', label,'.dta'))
}

get_year_bhps <- function(letter){
  return(1991 + (utf8ToInt(letter) - utf8ToInt('a')))
}

get_year_ukhls <- function(letter){
  return(2009 + (utf8ToInt(letter) - utf8ToInt('a')))
}
```

```{r}
read_youth_by_letter <- function(x, get_data_path, get_full_label, get_year){
  data <- read_stata(get_data_path('youth', x))
  data <- data %>% dplyr::select(any_of(c('pidp', 
                         get_full_label('hidp', x),
                         get_full_label('ypdoby', x),
                         get_full_label('ypsex', x),
                         get_full_label('ypnpal', x),
                         get_full_label('ypfpark', x),
                         get_full_label('ypmkfrn', x),
                         get_full_label('yppalo', x),
                         get_full_label('yppals', x),
                         get_full_label('ypsdqk', x)))) %>% 
          rename_with(~ 'birthy', all_of(get_full_label('ypdoby', x))) %>% 
          rename_with(~ 'female', all_of(get_full_label('ypsex', x))) %>%
          rename_with(~ 'hidp', all_of(get_full_label('hidp', x))) %>% 
          rename_with(~ case_when(. == get_full_label('ypnpal', x) ~ 'close_friends', 
                                  . == get_full_label('ypfpark', x) ~ 'outdoors_friends', 
                                  . == get_full_label('ypmkfrn', x) ~ 'easy_friends', 
                                  . == get_full_label('yppalo', x) ~ 'meeting_friends', 
                                  . == get_full_label('yppals', x) ~ 'in_house_friends',
                                  . == get_full_label('ypsdqk', x) ~ 'good_friend',
                                  TRUE ~ .))
  data$birthy = ifelse(data$birthy > 1000, data$birthy, data$birthy + 1900)
  data$year = get_year(x)
  data$age = data$year - data$birthy
  
  # Preparing percentile variables
  if('close_friends' %in% colnames(data)){
    data$close_friends <- ifelse(data$close_friends >= 0, data$close_friends, NaN)
    # 79 is mean + 3*sd of the entire sample, also seems like a quite reasonable,
    # conservative threshold - it's quite unlikely someone genuinely have 80 close friends
    data$close_friends <- ifelse(data$close_friends < 79, data$close_friends, NaN)
    data$p_close_friends <- 
      rank(data$close_friends, na.last = 'keep', ties.method = 'average')/
      sum(!is.na(data$close_friends))
  }
  
  if('outdoors_friends' %in% colnames(data)){
    data$outdoors_friends <- ifelse(data$outdoors_friends >= 0, data$outdoors_friends, NaN)
    data$p_outdoors_friends <- 
      rank(-data$outdoors_friends, na.last = 'keep', ties.method = 'average')/
      sum(!is.na(data$outdoors_friends))
  }
  
  if('easy_friends' %in% colnames(data)){
    data$easy_friends <- ifelse(data$easy_friends >= 0, data$easy_friends, NaN)
    data$p_easy_friends <- 
      rank(-data$easy_friends, na.last = 'keep', ties.method = 'average')/
      sum(!is.na(data$easy_friends))
  }
  
  if('meeting_friends' %in% colnames(data)){
    data$meeting_friends <- ifelse(data$meeting_friends >= 0, data$meeting_friends, NaN)
    data$p_meeting_friends <- 
      rank(data$meeting_friends, na.last = 'keep', ties.method = 'average')/
      sum(!is.na(data$meeting_friends))
  }

  if('in_house_friends' %in% colnames(data)){
    data$in_house_friends <- ifelse(data$in_house_friends >= 0, data$in_house_friends, NaN)
    data$p_in_house_friends <- 
      rank(data$in_house_friends, na.last = 'keep', ties.method = 'average')/
      sum(!is.na(data$in_house_friends))
  }
  
  if('good_friend' %in% colnames(data)){
    data$good_friend <- ifelse(data$good_friend >= 0, data$good_friend, NaN)
    data$p_good_friend <- 
      rank(data$good_friend, na.last = 'keep', ties.method = 'average')/
      sum(!is.na(data$good_friend))
  }
  
  data_household <- read_stata(get_data_path('hhresp', x))
  data_household <- data_household %>% dplyr::select(any_of(c(
                         get_full_label('hidp', x),
                         get_full_label('gor_dv', x),
                         get_full_label('hhsize', x),
                         get_full_label('fihhmngrs_dv', x)))) %>% 
          rename_with(~ 'hidp', all_of(get_full_label('hidp', x))) %>% 
          rename_with(~ 'gor', all_of(get_full_label('gor_dv', x))) %>% 
          rename_with(~ 'hh_income', all_of(get_full_label('fihhmngrs_dv', x))) %>%
          mutate(hh_income = ifelse((hh_income == -9) | (hh_income == -7), NA, hh_income))

  data_merged <- data %>% inner_join(data_household, by = c('hidp'))
  return(data_merged)
}

read_youth_bhps_by_letter <- function(x){
  return(read_youth_by_letter(x, get_data_path_bhps, get_full_label_bhps, get_year_bhps))
}

read_youth_ukhls_by_letter <- function(x){
  return(read_youth_by_letter(x, get_data_path_ukhls, get_full_label_ukhls, get_year_ukhls))
}

data_youth_bhps <- lapply(letters[4:18], read_youth_bhps_by_letter)
data_youth_ukhls <- lapply(letters[1:11], read_youth_ukhls_by_letter)
data_youth_merged <- rbindlist(c(data_youth_bhps, data_youth_ukhls), fill = T)

data_youth_merged <- filter(data_youth_merged, female > 0)
data_youth_merged$female <- data_youth_merged$female - 1
data_youth_merged <- filter(data_youth_merged, gor > 0)
data_youth_merged <- filter(data_youth_merged, birthy > 1970)
data_youth_merged <- filter(data_youth_merged, birthy < 2010)
data_youth_merged <- filter(data_youth_merged, age <= 18)

write.csv(data_youth_merged, 'data_youth_merged.csv')

# Variance covariance matrix
cor(data_youth_merged %>% 
  dplyr::select(c('p_close_friends', 'p_outdoors_friends', 'p_easy_friends', 'p_meeting_friends', 'p_in_house_friends')), use='pairwise.complete.obs')
```

```{r}
read_indresp_by_letter <- function(x, get_data_path, get_full_label, get_year){
  data <- read_stata(get_data_path('indresp', x))
  data <- data %>% dplyr::select(any_of(c('pidp',
                         get_full_label('hidp', x),
                         get_full_label('gor_dv', x),
                         get_full_label('urban_dv', x),
                         get_full_label('birthy', x),
                         get_full_label('sex', x),
                         get_full_label('plnowm', x),
                         get_full_label('memploy', x),
                         get_full_label('movjba', x),
                         get_full_label('family', x),
                         get_full_label('education', x),
                         get_full_label('housing', x),
                         get_full_label('area', x),
                         get_full_label('movy13', x),
                         get_full_label('fimnlabgrs_dv', x),
                         get_full_label('jbnssec_dv', x), 
                         get_full_label('casmin', x),
                         get_full_label('nchild_dv', x),
                         get_full_label('qfhigh_dv', x)))) %>%
          rename_with(~ 'hidp', all_of(get_full_label('hidp', x))) %>% 
          rename_with(~ 'gor', all_of(get_full_label('gor_dv', x))) %>%
          rename_with(~ 'birthy', all_of(get_full_label('birthy', x))) %>% 
          rename_with(~ 'female', all_of(get_full_label('sex', x))) %>%
          rename_with(~ 'moved', all_of(get_full_label('plnowm', x))) %>%
          rename_with(~ 'job_income', all_of(get_full_label('fimnlabgrs_dv', x))) %>%
          rename_with(~ 'job_type', all_of(get_full_label('jbnssec_dv', x))) %>%
          rename_with(~ 'nchild', all_of(get_full_label('nchild_dv', x))) %>% 
          rename_with(~ case_when(. == get_full_label('qfhigh_dv', x) ~ 'edu_new',
                                  . == get_full_label('casmin', x) ~ 'edu_old',
                                  . == get_full_label('urban_dv', x) ~ 'urban',
                                  . == get_full_label('memploy', x) ~ 'memploy',
                                  . == get_full_label('family', x) ~ 'family',
                                  . == get_full_label('education', x) ~ 'education',
                                  . == get_full_label('housing', x) ~ 'housing',
                                  . == get_full_label('area', x) ~ 'area',
                                  . == get_full_label('movjba', x) ~ 'memploy_old',
                                  . == get_full_label('movy13', x) ~ 'job_reason',
                                  TRUE ~ .)) %>%
            filter(is.na(job_income) | job_income >= 0)
  data$birthy = ifelse(data$birthy > 1000, data$birthy, data$birthy + 1900)
  data$year = get_year(x)
  data$age = data$year - data$birthy
  data$moved <- data$moved > 0
  data$employed = 1 - 1*(is.na(data$job_income) | data$job_income == 0)
  
  if('job_reason' %in% colnames(data)){
    data$job_reason <- as.numeric(data$job_reason == 0)
  }
  
  data_household <- read_stata(get_data_path('hhresp', x))
  data_household <- data_household %>% dplyr::select(any_of(c(
                         get_full_label('hidp', x),
                         get_full_label('hhsize', x),
                         get_full_label('fihhmngrs_dv', x)))) %>% 
          rename_with(~ 'hidp', all_of(get_full_label('hidp', x))) %>% 
          rename_with(~ 'hhsize', all_of(get_full_label('hhsize', x))) %>% 
          rename_with(~ 'hh_income', all_of(get_full_label('fihhmngrs_dv', x))) %>%
          mutate(hh_income = ifelse((hh_income == -9) | (hh_income == -7), NA, hh_income))
  
  data_merged <- data %>% inner_join(data_household, by = c('hidp'))
  return(data_merged)
}

read_indresp_bhps_by_letter <- function(x){
  return(read_indresp_by_letter(x, get_data_path_bhps, get_full_label_bhps, get_year_bhps))
}

read_indresp_ukhls_by_letter <- function(x){
  return(read_indresp_by_letter(x, get_data_path_ukhls, get_full_label_ukhls, get_year_ukhls))
}

data_adults_bhps <- lapply(letters[2:18], read_indresp_bhps_by_letter)
data_adults_ukhls <- lapply(letters[2:11], read_indresp_ukhls_by_letter)
data_adults_merged <- rbindlist(c(data_adults_bhps, data_adults_ukhls), fill = T)
```


Moving - data preparation
```{r}
data_adults_merged$moved_bin = 1*(data_adults_merged$moved == 1)
data_adults_merged$moved_for_job = ifelse(is.na(data_adults_merged$memploy), 
                                          1*(data_adults_merged$memploy_old > 0),
                                          1*(data_adults_merged$memploy > 0))
data_adults_merged$moved_to_look_for_job = 1*(data_adults_merged$memploy == 8)
data_adults_merged$moved_for_family = 1*(data_adults_merged$family > 0)
data_adults_merged$moved_for_education = 1*(data_adults_merged$education > 0)
data_adults_merged$moved_for_housing = 1*(data_adults_merged$housing > 0)
data_adults_merged$moved_for_area = 1*(data_adults_merged$area > 0)
data_adults_merged$moved_for_career = 1*(data_adults_merged$moved_for_job == 1 | 
                                           data_adults_merged$moved_for_education == 1)
data_adults_merged$moved_not_for_job = 
  data_adults_merged$moved_bin - data_adults_merged$moved_for_job
data_adults_merged$moved_not_for_career = 
  data_adults_merged$moved_bin - data_adults_merged$moved_for_career
data_adults_merged$moved_job_not_mentioned = 
  data_adults_merged$moved_bin - data_adults_merged$job_reason

data_adults_merged <- data_adults_merged %>%
    group_by(pidp) %>%
    arrange(year, .by_group = T) %>%
    filter(is.na(moved_not_for_job) | moved_not_for_job >= 0) %>%
    filter(is.na(moved_not_for_career) | moved_not_for_career >= 0) %>%
    filter(is.na(moved_job_not_mentioned) | moved_job_not_mentioned >= 0) %>%
    mutate(moved_times = cumsum(moved_bin)) %>%
    mutate(moved_total = sum(moved_bin)) %>%
    mutate(moved_for_job_times = cumsum(moved_for_job)) %>%
    mutate(moved_for_job_total = sum(moved_for_job)) %>%
    mutate(moved_not_for_job_times = cumsum(moved_not_for_job)) %>%
    mutate(moved_not_for_job_total = sum(moved_not_for_job)) %>%
    mutate(moved_not_for_career_times = cumsum(moved_not_for_career)) %>%
    mutate(moved_not_for_career_total = sum(moved_not_for_career)) %>%
    mutate(moved_job_not_mentioned_times = cumsum(moved_job_not_mentioned)) %>%
    mutate(moved_job_not_mentioned_total = sum(moved_job_not_mentioned)) %>%
    mutate(moved_to_look_for_job_times = cumsum(moved_to_look_for_job)) %>%
    mutate(moved_to_look_for_job_total = sum(moved_to_look_for_job)) %>%
    mutate(moved_for_family_times = cumsum(moved_for_family)) %>%
    mutate(moved_for_family_total = sum(moved_for_family)) %>%
    mutate(moved_for_education_times = cumsum(moved_for_education)) %>%
    mutate(moved_for_education_total = sum(moved_for_education)) %>%
    mutate(moved_for_housing_times = cumsum(moved_for_housing)) %>%
    mutate(moved_for_housing_total = sum(moved_for_housing)) %>%
    mutate(moved_for_area_times = cumsum(moved_for_area)) %>%
    mutate(moved_for_area_total = sum(moved_for_area)) %>%
    mutate(moved_for_career_times = cumsum(moved_for_career)) %>%
    mutate(moved_for_career_total = sum(moved_for_area)) %>%
    mutate(moved_not_for_job_lag_1 = shift(moved_not_for_job, 1, fill = 0, type = 'lag')) %>%
    mutate(moved_not_for_job_lag_2 = shift(moved_not_for_job, 2, fill = 0, type = 'lag')) %>%
    mutate(moved_not_for_job_lag_3 = shift(moved_not_for_job, 3, fill = 0, type = 'lag')) %>%
    mutate(moved_not_for_job_lag_4 = shift(moved_not_for_job, 4, fill = 0, type = 'lag')) %>%
    mutate(moved_not_for_job_lag_5 = shift(moved_not_for_job, 5, fill = 0, type = 'lag')) %>%
    mutate(moved_not_for_job_lead_1 = shift(moved_not_for_job, 1, fill = 0, type = 'lead')) %>%
    mutate(moved_not_for_job_lead_2 = shift(moved_not_for_job, 2, fill = 0, type = 'lead')) %>%
    mutate(moved_not_for_job_lead_3 = shift(moved_not_for_job, 3, fill = 0, type = 'lead')) %>%
    mutate(moved_not_for_job_lead_4 = shift(moved_not_for_job, 4, fill = 0, type = 'lead')) %>%
    mutate(moved_not_for_job_lead_5 = shift(moved_not_for_job, 5, fill = 0, type = 'lead')) %>%
    mutate(moved_not_for_career_lag_1 = shift(moved_not_for_career, 1, fill = 0, type = 'lag')) %>%
    mutate(moved_not_for_career_lag_2 = shift(moved_not_for_career, 2, fill = 0, type = 'lag')) %>%
    mutate(moved_not_for_career_lag_3 = shift(moved_not_for_career, 3, fill = 0, type = 'lag')) %>%
    mutate(moved_not_for_career_lag_4 = shift(moved_not_for_career, 4, fill = 0, type = 'lag')) %>%
    mutate(moved_not_for_career_lag_5 = shift(moved_not_for_career, 5, fill = 0, type = 'lag')) %>%
    mutate(moved_not_for_career_lead_1 = 
             shift(moved_not_for_career, 1, fill = 0, type = 'lead')) %>%
    mutate(moved_not_for_career_lead_2 = 
             shift(moved_not_for_career, 2, fill = 0, type = 'lead')) %>%
    mutate(moved_not_for_career_lead_3 = 
             shift(moved_not_for_career, 3, fill = 0, type = 'lead')) %>%
    mutate(moved_not_for_career_lead_4 = 
             shift(moved_not_for_career, 4, fill = 0, type = 'lead')) %>%
    mutate(moved_not_for_career_lead_5 = 
             shift(moved_not_for_career, 5, fill = 0, type = 'lead')) %>%
    mutate(moved_for_education_lag_1 = shift(moved_for_education, 1, fill = 0, type = 'lag')) %>%
    mutate(moved_for_education_lag_2 = shift(moved_for_education, 2, fill = 0, type = 'lag')) %>%
    mutate(moved_for_education_lag_3 = shift(moved_for_education, 3, fill = 0, type = 'lag')) %>%
    mutate(moved_for_education_lag_4 = shift(moved_for_education, 4, fill = 0, type = 'lag')) %>%
    mutate(moved_for_education_lag_5 = shift(moved_for_education, 5, fill = 0, type = 'lag')) %>%
    mutate(moved_for_education_lead_1 = shift(moved_for_education, 1, fill = 0, type = 'lead')) %>%
    mutate(moved_for_education_lead_2 = shift(moved_for_education, 2, fill = 0, type = 'lead')) %>%
    mutate(moved_for_education_lead_3 = shift(moved_for_education, 3, fill = 0, type = 'lead')) %>%
    mutate(moved_for_education_lead_4 = shift(moved_for_education, 4, fill = 0, type = 'lead')) %>%
    mutate(moved_for_education_lead_5 = shift(moved_for_education, 5, fill = 0, type = 'lead')) %>%
    mutate(moved_for_family_lag_1 = shift(moved_for_family, 1, fill = 0, type = 'lag')) %>%
    mutate(moved_for_family_lag_2 = shift(moved_for_family, 2, fill = 0, type = 'lag')) %>%
    mutate(moved_for_family_lag_3 = shift(moved_for_family, 3, fill = 0, type = 'lag')) %>%
    mutate(moved_for_family_lag_4 = shift(moved_for_family, 4, fill = 0, type = 'lag')) %>%
    mutate(moved_for_family_lag_5 = shift(moved_for_family, 5, fill = 0, type = 'lag')) %>%
    mutate(moved_for_family_lead_1 = shift(moved_for_family, 1, fill = 0, type = 'lead')) %>%
    mutate(moved_for_family_lead_2 = shift(moved_for_family, 2, fill = 0, type = 'lead')) %>%
    mutate(moved_for_family_lead_3 = shift(moved_for_family, 3, fill = 0, type = 'lead')) %>%
    mutate(moved_for_family_lead_4 = shift(moved_for_family, 4, fill = 0, type = 'lead')) %>%
    mutate(moved_for_family_lead_5 = shift(moved_for_family, 5, fill = 0, type = 'lead')) %>%
    mutate(moved_for_housing_lag_1 = shift(moved_for_housing, 1, fill = 0, type = 'lag')) %>%
    mutate(moved_for_housing_lag_2 = shift(moved_for_housing, 2, fill = 0, type = 'lag')) %>%
    mutate(moved_for_housing_lag_3 = shift(moved_for_housing, 3, fill = 0, type = 'lag')) %>%
    mutate(moved_for_housing_lag_4 = shift(moved_for_housing, 4, fill = 0, type = 'lag')) %>%
    mutate(moved_for_housing_lag_5 = shift(moved_for_housing, 5, fill = 0, type = 'lag')) %>%
    mutate(moved_for_housing_lead_1 = shift(moved_for_housing, 1, fill = 0, type = 'lead')) %>%
    mutate(moved_for_housing_lead_2 = shift(moved_for_housing, 2, fill = 0, type = 'lead')) %>%
    mutate(moved_for_housing_lead_3 = shift(moved_for_housing, 3, fill = 0, type = 'lead')) %>%
    mutate(moved_for_housing_lead_4 = shift(moved_for_housing, 4, fill = 0, type = 'lead')) %>%
    mutate(moved_for_housing_lead_5 = shift(moved_for_housing, 5, fill = 0, type = 'lead')) %>%
    mutate(moved_for_area_lag_1 = shift(moved_for_area, 1, fill = 0, type = 'lag')) %>%
    mutate(moved_for_area_lag_2 = shift(moved_for_area, 2, fill = 0, type = 'lag')) %>%
    mutate(moved_for_area_lag_3 = shift(moved_for_area, 3, fill = 0, type = 'lag')) %>%
    mutate(moved_for_area_lag_4 = shift(moved_for_area, 4, fill = 0, type = 'lag')) %>%
    mutate(moved_for_area_lag_5 = shift(moved_for_area, 5, fill = 0, type = 'lag')) %>%
    mutate(moved_for_area_lead_1 = shift(moved_for_area, 1, fill = 0, type = 'lead')) %>%
    mutate(moved_for_area_lead_2 = shift(moved_for_area, 2, fill = 0, type = 'lead')) %>%
    mutate(moved_for_area_lead_3 = shift(moved_for_area, 3, fill = 0, type = 'lead')) %>%
    mutate(moved_for_area_lead_4 = shift(moved_for_area, 4, fill = 0, type = 'lead')) %>%
    mutate(moved_for_area_lead_5 = shift(moved_for_area, 5, fill = 0, type = 'lead'))
  
for(i in 1:max(data_adults_merged$moved_not_for_job_times, na.rm = T)){
  data_adults_merged <- data_adults_merged %>%
    group_by(pidp) %>%
    arrange(year, .by_group = T) %>%
    mutate("moved_not_for_job_bin_lag_{i}" := 
             1*(shift(moved_not_for_job_times, 6, fill = 0, type = 'lag') >= i)) %>%
    mutate("moved_not_for_job_bin_lead_{i}" := 
             1*(shift(moved_not_for_job_times, 6, fill = 0, type = 'lead') <= i & 
                  moved_not_for_job_total >= i))
}

for(i in 1:max(data_adults_merged$moved_not_for_career_times, na.rm = T)){
  data_adults_merged <- data_adults_merged %>%
    group_by(pidp) %>%
    arrange(year, .by_group = T) %>%
    mutate("moved_not_for_career_bin_lag_{i}" := 
             1*(shift(moved_not_for_career_times, 6, fill = 0, type = 'lag') >= i)) %>%
    mutate("moved_not_for_career_bin_lead_{i}" := 
             1*(shift(moved_not_for_career_times, 6, fill = 0, type = 'lead') <= i & 
                  moved_not_for_career_total >= i))
}
    
for(i in 1:max(data_adults_merged$moved_for_education_times, na.rm = T)){
  data_adults_merged <- data_adults_merged %>%
    group_by(pidp) %>%
    arrange(year, .by_group = T) %>%    
    mutate("moved_for_education_bin_lag_{i}" := 
             1*(shift(moved_for_education_times, 6, fill = 0, type = 'lag') >= i)) %>%
    mutate("moved_for_education_bin_lead_{i}" := 
             1*(shift(moved_for_education_times, 6, fill = 0, type = 'lead') <= i & 
                  moved_for_education_total >= i)) 
}

for(i in 1:max(data_adults_merged$moved_for_family_times, na.rm = T)){
  data_adults_merged <- data_adults_merged %>%
    group_by(pidp) %>%
    arrange(year, .by_group = T) %>%    
    mutate("moved_for_family_bin_lag_{i}" := 
             1*(shift(moved_for_family_times, 6, fill = 0, type = 'lag') >= i)) %>%
    mutate("moved_for_family_bin_lead_{i}" := 
             1*(shift(moved_for_family_times, 6, fill = 0, type = 'lead') <= i & 
                  moved_for_family_total >= i))
}

for(i in 1:max(data_adults_merged$moved_for_housing_times, na.rm = T)){
  data_adults_merged <- data_adults_merged %>%
    group_by(pidp) %>%
    arrange(year, .by_group = T) %>%  
    mutate("moved_for_housing_bin_lag_{i}" := 
             1*(shift(moved_for_housing_times, 6, fill = 0, type = 'lag') >= i)) %>%
    mutate("moved_for_housing_bin_lead_{i}" := 
             1*(shift(moved_for_housing_times, 6, fill = 0, type = 'lead') <= i & 
                  moved_for_housing_total >= i))
}

for(i in 1:max(data_adults_merged$moved_for_area_times, na.rm = T)){
  data_adults_merged <- data_adults_merged %>%
    group_by(pidp) %>%
    arrange(year, .by_group = T) %>%  
    mutate("moved_for_area_bin_lag_{i}" := 
             1*(shift(moved_for_area_times, 6, fill = 0, type = 'lag') >= i)) %>%
    mutate("moved_for_area_bin_lead_{i}" := 
             1*(shift(moved_for_area_times, 6, fill = 0, type = 'lead') <= i & 
                  moved_for_area_total >= i))
}

data_adults_merged$urban <- ifelse(data_adults_merged$urban == -9, NA, data_adults_merged$urban)
data_adults_merged$urban <- -(data_adults_merged$urban - 2)
data_adults_merged <- filter(data_adults_merged, female > 0)
data_adults_merged$female <- data_adults_merged$female - 1
data_adults_merged <- filter(data_adults_merged, birthy > 1890)
data_adults_merged <- filter(data_adults_merged, birthy < 2010)
data_adults_merged <- filter(data_adults_merged, age < 100)
data_adults_merged <- filter(data_adults_merged, age > 18)
data_adults_merged <- filter(data_adults_merged, gor > 0)
data_adults_merged <- filter(data_adults_merged, job_type != -9)
data_adults_merged <- filter(data_adults_merged, job_type != -1)
data_adults_merged$education_level = 0

data_adults_merged$education_level = ifelse((data_adults_merged$edu_old == -9) %in% TRUE |
                                              (data_adults_merged$edu_new == -9) %in% TRUE,
                                            -9, data_adults_merged$education_level)
data_adults_merged$education_level = ifelse((data_adults_merged$edu_old == -8) %in% TRUE |
                                              (data_adults_merged$edu_new == -8) %in% TRUE,
                                            -8, data_adults_merged$education_level)
data_adults_merged$education_level = ifelse((data_adults_merged$edu_old == -7) %in% TRUE |
                                              (data_adults_merged$edu_new == -7) %in% TRUE,
                                            -7, data_adults_merged$education_level)
# Higer ed
data_adults_merged$education_level = ifelse((data_adults_merged$edu_old == 9) %in% TRUE |
                                              (data_adults_merged$edu_new == 1) %in% TRUE,
                                            1, data_adults_merged$education_level)

# lower tert
data_adults_merged$education_level = ifelse((data_adults_merged$edu_old == 8) %in% TRUE |
                                              (data_adults_merged$edu_new == 2) %in% TRUE | 
                                              (data_adults_merged$edu_new == 3) %in% TRUE | 
                                              (data_adults_merged$edu_new == 4) %in% TRUE | 
                                              (data_adults_merged$edu_new == 5) %in% TRUE,
                                            2, data_adults_merged$education_level)

# middle ed
data_adults_merged$education_level = ifelse((data_adults_merged$edu_old == 4) %in% TRUE |
                                              (data_adults_merged$edu_old == 5) %in% TRUE |
                                              (data_adults_merged$edu_old == 6) %in% TRUE |
                                              (data_adults_merged$edu_old == 7) %in% TRUE |
                                              (data_adults_merged$edu_new == 7) %in% TRUE | 
                                              (data_adults_merged$edu_new == 8) %in% TRUE | 
                                              (data_adults_merged$edu_new == 9) %in% TRUE | 
                                              (data_adults_merged$edu_new == 10) %in% TRUE | 
                                              (data_adults_merged$edu_new == 11) %in% TRUE | 
                                              (data_adults_merged$edu_new == 12) %in% TRUE | 
                                              (data_adults_merged$edu_new == 13) %in% TRUE,
                                            3, data_adults_merged$education_level)

# lower ed
data_adults_merged$education_level = ifelse((data_adults_merged$edu_old == 1) %in% TRUE |
                                              (data_adults_merged$edu_old == 2) %in% TRUE |
                                              (data_adults_merged$edu_old == 3) %in% TRUE |
                                              (data_adults_merged$edu_new == 14) %in% TRUE | 
                                              (data_adults_merged$edu_new == 15) %in% TRUE | 
                                              (data_adults_merged$edu_new == 16) %in% TRUE,
                                            4, data_adults_merged$education_level)

# other
data_adults_merged$education_level = ifelse((data_adults_merged$edu_new == 96) %in% TRUE,
                                            5, data_adults_merged$education_level)

data_adults_merged <- filter(data_adults_merged, education_level != 0)
write.csv(data_adults_merged, 'data_adults_merged.csv')
```
